SkipFirstRow is configurable with pipeline
