+++ enhydrator/src/main/java/com/airhacks/enhydrator/flexpipe/Plumber.java
@@ -25,10 +25,11 @@ import com.airhacks.enhydrator.in.JDBCSource;
 import com.airhacks.enhydrator.in.VirtualSinkSource;
 import com.airhacks.enhydrator.out.JDBCSink;
 import com.airhacks.enhydrator.out.LogSink;
-import com.airhacks.enhydrator.transform.DatatypeMapper;
 import com.airhacks.enhydrator.transform.Datatype;
+import com.airhacks.enhydrator.transform.DatatypeMapper;
 import com.airhacks.enhydrator.transform.DestinationMapper;
 import com.airhacks.enhydrator.transform.NashornRowTransformer;
+import com.airhacks.enhydrator.transform.SkipFirstRow;
 import com.airhacks.enhydrator.transform.TargetMapping;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
@@ -72,7 +73,7 @@ public class Plumber {
                     Pipeline.class, JDBCSink.class, LogSink.class,
                     UnmanagedConnectionProvider.class, ColumnTransformation.class,
                     NashornRowTransformer.class, DestinationMapper.class,
-                    TargetMapping.class, DatatypeMapper.class, Datatype.class);
+                    TargetMapping.class, DatatypeMapper.class, Datatype.class, SkipFirstRow.class);
             this.marshaller = context.createMarshaller();
             this.marshaller.setProperty(Marshaller.JAXB_FORMATTED_OUTPUT, true);
             this.unmarshaller = context.createUnmarshaller();
