--- enhydrator/src/main/java/com/airhacks/enhydrator/functions/SkipFirstRow.java
@@ -1,44 +1,0 @@
-package com.airhacks.enhydrator.functions;
-
-/*
- * #%L
- * enhydrator
- * %%
- * Copyright (C) 2014 Adam Bien
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-import com.airhacks.enhydrator.in.Row;
-import com.airhacks.enhydrator.transform.RowTransformer;
-
-/**
- *
- * @author airhacks.com
- */
-public class SkipFirstRow implements RowTransformer {
-
-    private boolean skipped = false;
-
-    @Override
-    public Row execute(Row input) {
-        if (skipped) {
-            return input;
-        } else {
-            skipped = true;
-            return null;
-        }
-
-    }
-
-}
