+++ enhydrator/src/main/java/com/airhacks/enhydrator/transform/SkipFirstRow.java
@@ -1,0 +1,73 @@
+package com.airhacks.enhydrator.transform;
+
+/*
+ * #%L
+ * enhydrator
+ * %%
+ * Copyright (C) 2014 Adam Bien
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+import com.airhacks.enhydrator.flexpipe.RowTransformation;
+import com.airhacks.enhydrator.in.Row;
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlRootElement;
+import javax.xml.bind.annotation.XmlTransient;
+
+/**
+ *
+ * @author airhacks.com
+ */
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlRootElement(name = "skip-first-row")
+public class SkipFirstRow extends RowTransformation {
+
+    @XmlTransient
+    private boolean skipped = false;
+
+    @Override
+    public Row execute(Row input) {
+        if (skipped) {
+            return input;
+        } else {
+            skipped = true;
+            return null;
+        }
+
+    }
+
+    @Override
+    public int hashCode() {
+        int hash = 7;
+        hash = 67 * hash + (this.skipped ? 1 : 0);
+        return hash;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj == null) {
+            return false;
+        }
+        if (getClass() != obj.getClass()) {
+            return false;
+        }
+        final SkipFirstRow other = (SkipFirstRow) obj;
+        if (this.skipped != other.skipped) {
+            return false;
+        }
+        return true;
+    }
+
+}
